+++ Source/ImageLoader.h
@@ -40,12 +40,12 @@ class ImageLoader final
 public:
     struct ResultInfo
     {
-        uint32_t       levelOffsets[ MAX_PREGENERATED_MIPMAP_LEVELS ];
-        uint32_t       levelSizes[ MAX_PREGENERATED_MIPMAP_LEVELS ];
+        size_t         levelOffsets[ MAX_PREGENERATED_MIPMAP_LEVELS ];
+        size_t         levelSizes[ MAX_PREGENERATED_MIPMAP_LEVELS ];
         uint32_t       levelCount;
         bool           isPregenerated;
         const uint8_t* pData;
-        uint32_t       dataSize;
+        size_t         dataSize;
         RgExtent2D     baseSize;
         VkFormat       format;
     };
@@ -53,7 +53,7 @@ public:
     struct LayeredResultInfo
     {
         std::vector< const uint8_t* > layerData;
-        uint32_t                      dataSize;
+        size_t                        dataSize;
         RgExtent2D                    baseSize;
         VkFormat                      format;
     };
